# perf(ext/request): optimize validate and normalize HTTP method